Removed browser_unit implicit wait
As described in the documentation "An implicit wait is to tell WebDriver
to poll the DOM for a certain amount of time when trying to find an
element or elements if they are not immediately available."  In our
use case of searching for ads, this causes an unacceptable performance
impact.

Especially when recursing through iframes this would cause orders of
magnitude slowdown.  If the elements are not available that is ok, we
can safely not check them.

On the basis of a small selection of sites this does not appear to have
a negative impact on the number of ads collected.  Additionally as this
setting "is set for the life of the WebDriver object instance." the
impact will slow down any other drivers as well.